hauteur label ok + tranduction fr ok
