+++ interface/translations/__init__.py
@@ -1,5 +1,8 @@
-from .Fr_fr import FR_fr
+from .fr_FR import fr_FR
 
-__all__:[
-    "FR_fr",
-]
+TRANSLATIONS = {
+    "fr_FR": fr_FR,
+    # "en_US": en_US,
+}
+
+__all__ = ["TRANSLATIONS"]
